Initializer moved to mixin and tests updated accordingly
