--- app/initializers/add-tooltips-to-views.js
@@ -1,1 +1,0 @@
-export { default, initialize } from 'ember-tooltips/initializers/add-tooltips-to-views';
