+++ app/mixins/components/tooltips.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-tooltips/mixins/components/tooltips';
