+++ tests/unit/initializers/add-tooltips-to-views.js
